refactor(api): move usecase to team context
