+++ api/tests/team/integration/domain/usecases/update-membership.usecase.test.js
@@ -1,7 +1,7 @@
-import { updateMembership } from '../../../../lib/domain/usecases/update-membership.js';
-import * as membershipRepository from '../../../../lib/infrastructure/repositories/membership-repository.js';
-import { Membership } from '../../../../src/shared/domain/models/Membership.js';
-import { databaseBuilder, expect } from '../../../test-helper.js';
+import * as membershipRepository from '../../../../../lib/infrastructure/repositories/membership-repository.js';
+import { Membership } from '../../../../../src/shared/domain/models/Membership.js';
+import { updateMembership } from '../../../../../src/team/domain/usecases/update-membership.usecase.js';
+import { databaseBuilder, expect } from '../../../../test-helper.js';
 
 describe('Integration | UseCases | update-membership', function () {
   it('should update membership', async function () {
