+++ api/tests/team/unit/domain/usecases/update-membership.usecase.test.js
@@ -1,7 +1,7 @@
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { InvalidMembershipOrganizationRoleError } from '../../../../src/shared/domain/errors.js';
-import { Membership } from '../../../../src/shared/domain/models/Membership.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { InvalidMembershipOrganizationRoleError } from '../../../../../src/shared/domain/errors.js';
+import { Membership } from '../../../../../src/shared/domain/models/Membership.js';
+import { usecases } from '../../../../../src/team/domain/usecases/index.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 const { updateMembership } = usecases;
